Remove parameters from vim_findfile_init
The only two places vim_findfile_init is called in the codebase are in
the convenience function find_file_in_path_option, and in tags.c. Both
of those places used 3 arguments that were the same, so they were
removed as arguments.

char_u *stopdirs: This argument was always created with the
vim_findfile_stopdir function, so we now do that inside of
vim_findfile_init.

int level: Sets the maximum number of levels to recurse into during
vim_findfile. Was always 100, and is now set in file_search.h

int free_visited: This determines whether or not to reuse old visited
lists. Was always FALSE, so that removes one branch in
vim_findfile_init.