Refactor search_with_suffixes, file_of_type_exists
The function search_with_suffixes  searches by trying to add a list of
suffixes to a file name one at a time. Also, this commit unrolls a for
loop that happened either once or twice based on a conditional.

The function file_of_type_exists is a conditional check that the file
exists, and is of the type that we are looking for.

Also, favor early return over goto, removing the need for the
"file_name == NULL" check for the errors